Refactor: Rename LoginActivity to MainActivity and update related files
This commit renames `LoginActivity` to `MainActivity`, updates the layout file name accordingly, and adjusts the `AndroidManifest.xml` and `SignupActivity` to use the new `MainActivity` name. It also adds new string resources for "password", "login", "sign up" and "username", and uses them in layout file. The `vcs.xml` file is updated to map the project to Git.